+++ src/OBD/data/BitEncoded/MonitorStatus.h
@@ -21,13 +21,13 @@ private:
      * Off or On, indicates if the CEL/MIL is on (or should be on)
      * * Byte A7
      */
-    unique_ptr<DataObject<bool>> mil = make_unique<DataObject<bool>>(A, (byte) 7);
+    unique_ptr<DataObject<bool>> mil = make_unique<DataObject<bool>>(A, 7);
 
     /**
      * Number of confirmed emissions-related DTCs available for display.
      * Byte A6-A0
      */
-    unique_ptr<DataObject<unsigned int>> dtcCount = make_unique<DataObject<unsigned int>>(A, (byte) 6, A, (byte) 0);
+    unique_ptr<DataObject<unsigned int>> dtcCount = make_unique<DataObject<unsigned int>>(A, 6, A, 0);
 
 
     /**
@@ -36,9 +36,9 @@ private:
     shared_ptr<Engine> engine;
 
 
-    unique_ptr<OBDTest> components = make_unique<OBDTest>("Components", B, (byte) 2, B, (byte) 6);
-    unique_ptr<OBDTest> fuelSystem = make_unique<OBDTest>("FuelSystem", B, (byte) 1, B, (byte) 5);
-    unique_ptr<OBDTest> misfire = make_unique<OBDTest>("Misfire", B, (byte) 0, B, (byte) 4);
+    unique_ptr<OBDTest> components = make_unique<OBDTest>("Components", B, 2, B, 6);
+    unique_ptr<OBDTest> fuelSystem = make_unique<OBDTest>("FuelSystem", B, 1, B, 5);
+    unique_ptr<OBDTest> misfire = make_unique<OBDTest>("Misfire", B, 0, B, 4);
 
 public:
     explicit MonitorStatus(shared_ptr<Engine> engine);
