+++ tests/src/OBD/data/DataObjectTest.cpp
@@ -49,65 +49,65 @@ void compareResponse(vector<byte> expectedResponse, DataObject<unsigned short> *
 
 
 TEST(DataObjectTest, DecodeBoolA0) {
-    auto *obj = new DataObject<bool>(A, (byte) 0);
+    auto *obj = new DataObject<bool>(A, 0);
     vector<byte> data{(byte) 0x01, (byte) 0xff, (byte) 0xff, (byte) 0x00};
     testBool(obj, data, true);
 }
 
 
 TEST(DataObjectTest, DecodeBoolA0_Negative) {
-    auto *obj = new DataObject<bool>(A, (byte) 0);
+    auto *obj = new DataObject<bool>(A, 0);
     vector<byte> data{(byte) 0xfe, (byte) 0x00, (byte) 0x00, (byte) 0x00};
     testBool(obj, data, false);
 }
 
 TEST(DataObjectTest, DecodeBoolA4) {
-    auto *obj = new DataObject<bool>(A, (byte) 4);
+    auto *obj = new DataObject<bool>(A, 4);
     vector<byte> data{(byte) 0x10, (byte) 0x00, (byte) 0x00, (byte) 0x00};
     testBool(obj, data, true);
 }
 
 
 TEST(DataObjectTest, DecodeBoolA7) {
-    auto *obj = new DataObject<bool>(A, (byte) 7);
+    auto *obj = new DataObject<bool>(A, 7);
     vector<byte> data{(byte) 0x80, (byte) 0x00, (byte) 0x00, (byte) 0x00};
     testBool(obj, data, true);
 }
 
 
 TEST(DataObjectTest, DecodeBoolC2) {
-    auto *obj = new DataObject<bool>(C, (byte) 2);
+    auto *obj = new DataObject<bool>(C, 2);
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0x04, (byte) 0x00};
     testBool(obj, data, true);
 }
 
 TEST(DataObjectTest, DecodeBoolC3) {
-    auto *obj = new DataObject<bool>(C, (byte) 4);
+    auto *obj = new DataObject<bool>(C, 4);
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0x10, (byte) 0x00};
     testBool(obj, data, true);
 }
 
 TEST(DataObjectTest, DecodeBoolC6) {
-    auto *obj = new DataObject<bool>(C, (byte) 6);
+    auto *obj = new DataObject<bool>(C, 6);
     // all 1 but 7 and 5
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0x5f, (byte) 0x00};
     testBool(obj, data, true);
 }
 
 TEST(DataObjectTest, DecodeBoolD7) {
-    auto *obj = new DataObject<bool>(D, (byte) 7);
+    auto *obj = new DataObject<bool>(D, 7);
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x80};
     testBool(obj, data, true);
 }
 
 TEST(DataObjectTest, DecodeBoolD7Negative) {
-    auto *obj = new DataObject<bool>(D, (byte) 7);
+    auto *obj = new DataObject<bool>(D, 7);
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x7f};
     testBool(obj, data, false);
 }
 
 TEST(DataObjectTest, DecodeBoolD2_Negative) {
-    auto *obj = new DataObject<bool>(D, (byte) 2);
+    auto *obj = new DataObject<bool>(D, 2);
     // every bit expect D2 is true
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0xFB};
     testBool(obj, data, false);
@@ -115,40 +115,40 @@ TEST(DataObjectTest, DecodeBoolD2_Negative) {
 
 
 TEST(DataObjectTest, DecodeBoolC7) {
-    auto *obj = new DataObject<bool>(C, (byte) 7, C, (byte) 7);
+    auto *obj = new DataObject<bool>(C, 7, C, 7);
     vector<byte> data{(byte) 0x01, (byte) 0xff, (byte) 0xff, (byte) 0xff};
     testBool(obj, data, true);
 }
 
 
 TEST(DataObjectTest, TestDecodeByteA) {
-    auto *obj = new DataObject<byte>(A, (byte) 7, A, (byte) 0);
+    auto *obj = new DataObject<byte>(A, 7, A, 0);
     vector<byte> data{(byte) 0x42, (byte) 0xff, (byte) 0xff, (byte) 0xff};
     byteTest(obj, data);
 }
 
 TEST(DataObjectTest, TestDecodeByteB) {
-    auto *obj = new DataObject<byte>(B, (byte) 7, B, (byte) 0);
+    auto *obj = new DataObject<byte>(B, 7, B, 0);
     vector<byte> data{(byte) 0xff, (byte) 0x42, (byte) 0xff, (byte) 0xff};
     byteTest(obj, data);
 }
 
 
 TEST(DataObjectTest, TestDecodeByteC) {
-    auto *obj = new DataObject<byte>(C, (byte) 7, C, (byte) 0);
+    auto *obj = new DataObject<byte>(C, 7, C, 0);
     vector<byte> data{(byte) 0xff, (byte) 0xff, (byte) 0x42, (byte) 0xff};
     byteTest(obj, data);
 }
 
 
 TEST(DataObjectTest, TestDecodeByteD) {
-    auto *obj = new DataObject<byte>(D, (byte) 7, D, (byte) 0);
+    auto *obj = new DataObject<byte>(D, 7, D, 0);
     vector<byte> data{(byte) 0xff, (byte) 0xff, (byte) 0xff, (byte) 0x42};
     byteTest(obj, data);
 }
 
 TEST(DataObjectTest, TestDecodeShortLeft) {
-    auto *obj = new DataObject<unsigned short>(A, (byte) 7, B, (byte) 0);
+    auto *obj = new DataObject<unsigned short>(A, 7, B, 0);
     vector<byte> data{(byte) 0xca, (byte) 0xfe};
     obj->fromFrame(data.data(), data.size());
     EXPECT_EQ(0xcafe, obj->getValue());
@@ -157,7 +157,7 @@ TEST(DataObjectTest, TestDecodeShortLeft) {
 
 
 TEST(DataObjectTest, TestDecodeShortMiddle) {
-    auto *obj = new DataObject<unsigned short>(B, (byte) 7, C, (byte) 0);
+    auto *obj = new DataObject<unsigned short>(B, 7, C, 0);
     vector<byte> data{(byte) 0x00, (byte) 0xca, (byte) 0xfe, (byte) 0x00};
     EXPECT_EQ(1, obj->fromFrame(data.data(), data.size()));
     EXPECT_EQ(0xcafe, obj->getValue());
@@ -166,7 +166,7 @@ TEST(DataObjectTest, TestDecodeShortMiddle) {
 
 
 TEST(DataObjectTest, TestDecodeShortRight) {
-    auto *obj = new DataObject<unsigned short>(C, (byte) 7, D, (byte) 0);
+    auto *obj = new DataObject<unsigned short>(C, 7, D, 0);
     vector<byte> data{(byte) 0x00, (byte) 0x00, (byte) 0xca, (byte) 0xfe};
     EXPECT_EQ(1, obj->fromFrame(data.data(), data.size()));
     EXPECT_EQ(0xcafe, obj->getValue());
@@ -174,7 +174,7 @@ TEST(DataObjectTest, TestDecodeShortRight) {
 }
 
 TEST(DataObjectTest, TestDecodeShortOdd) {
-    auto *obj = new DataObject<unsigned short>(A, (byte) 6, A, (byte) 0);
+    auto *obj = new DataObject<unsigned short>(A, 6, A, 0);
     vector<byte> data{(byte) 0xf1, (byte) 0x00, (byte) 0x00, (byte) 0x00};
     EXPECT_EQ(1, obj->fromFrame(data.data(), data.size()));
     EXPECT_EQ(0x71, obj->getValue());
@@ -184,7 +184,7 @@ TEST(DataObjectTest, TestDecodeShortOdd) {
 }
 
 TEST(DataObjectTest, TestDecodeShortOverlap) {
-    auto *obj = new DataObject<unsigned short>(A, (byte) 1, B, (byte) 6);
+    auto *obj = new DataObject<unsigned short>(A, 1, B, 6);
     vector<byte> data{(byte) 0x3c, (byte) 0x00, (byte) 0x00, (byte) 0x00};
     obj->fromFrame(data.data(), data.size());
     EXPECT_EQ(0x0f, obj->getValue());
@@ -197,7 +197,7 @@ TEST(DataObjectTest, TestDecodeShortOverlap) {
 TEST(DataObjectTest, TestDecode2BitAsShortBorder) {
     // A7	A6	A5	A4	A3	A2	A1	A0	B7
     // 00  01  02  03  04  05  06  07  08
-    auto *obj = new DataObject<unsigned short>(A, (byte) 0, B, (byte) 7);
+    auto *obj = new DataObject<unsigned short>(A, 0, B, 7);
 
     // data in binary
     // 0000 0001 1000 0000
@@ -211,7 +211,7 @@ TEST(DataObjectTest, TestDecode2BitAsShortBorder) {
 TEST(DataObjectTest, TestDecode2BitAsShortMiddle) {
     // A7	A6	A5	A4	A3	A2	A1	A0	B7
     // 00  01  02  03  04  05  06  07  08
-    auto *obj = new DataObject<unsigned short>(A, (byte) 5, A, (byte) 4);
+    auto *obj = new DataObject<unsigned short>(A, 5, A, 4);
 
     // data in binary
     // 7654 3210
@@ -226,7 +226,7 @@ TEST(DataObjectTest, TestDecode2BitAsShortMiddle) {
 TEST(DataObjectTest, TestDecode2BitAsShortMiddle2) {
     // A7	A6	A5	A4	A3	A2	A1	A0	B7
     // 00  01  02  03  04  05  06  07  08
-    auto *obj = new DataObject<unsigned short>(A, (byte) 3, A, (byte) 2);
+    auto *obj = new DataObject<unsigned short>(A, 3, A, 2);
 
     // data in binary
     // 7654 3210
@@ -241,7 +241,7 @@ TEST(DataObjectTest, TestDecode2BitAsShortMiddle2) {
 TEST(DataObjectTest, TestDecode3BitAsShortMiddle2) {
     // A7	A6	A5	A4	A3	A2	A1	A0	B7
     // 00  01  02  03  04  05  06  07  08
-    auto *obj = new DataObject<unsigned short>(A, (byte) 4, A, (byte) 2);
+    auto *obj = new DataObject<unsigned short>(A, 4, A, 2);
 
     // data in binary
     // 7654 3210
@@ -255,7 +255,7 @@ TEST(DataObjectTest, TestDecode3BitAsShortMiddle2) {
 TEST(DataObjectTest, TestDecode2BitMSB) {
     // A7	A6	A5	A4	A3	A2	A1	A0	B7
     // 00  01  02  03  04  05  06  07  08
-    auto *obj = new DataObject<unsigned short>(A, (byte) 7, A, (byte) 6);
+    auto *obj = new DataObject<unsigned short>(A, 7, A, 6);
 
     vector<byte> data{(byte) 0xC0};
     obj->fromFrame(data.data(), data.size());
@@ -266,7 +266,7 @@ TEST(DataObjectTest, TestDecode2BitMSB) {
 
 
 TEST(DataObjectTest, TestDecodeInt) {
-    auto *obj = new DataObject<unsigned int>(A, (byte) 7, D, (byte) 0);
+    auto *obj = new DataObject<unsigned int>(A, 7, D, 0);
     vector<byte> data{(byte) 0xca, (byte) 0xfe, (byte) 0xba, (byte) 0xbe,};
     EXPECT_EQ(1, obj->fromFrame(data.data(), data.size()));
     EXPECT_EQ(0xcafebabe, obj->getValue());
